refactor: extract `utils.nostr`
